TheObserverPattern
This version has code related to the Observer Pattern. The application is a Weather Station which takes temperature, humidity and pressure and displays the current data, statistics, forecast based on pressure data and heat index. 

What happens if user asks for Precipitation data? Currently in the setMeasurement method all the data are passed as argument. If I need to display rainfall then I need to change the setMeasurement and Update methods for all the observers. If client wants more displays which displays data corresponding to temp, pressure and humidity then it's easy. Just create one more observer and register it.